Doing block search by hitting the Postgres db.
Starting to rework the "Block.objects.search(...)" django formalism by rewriting the code to use the underlying openblock tables on the Postgis database directly.  Right now this is simply assembling queries by hand -- to it's not safe to use at all. We should rewrite this using some non-Django middle layer.