Added PostgisIntersectionSearch
Replicates the functionality of IntersectionManager.  This, along with PostgisBlockSearch, should be the two basic interfaces to Postgis that we need for geocoding.

Also, fixed a few bugs in PostgisBlockSearch, which would have caused it to break for some inputs.  (Mainly, fixed parameter names that I had copied-and-pasted but not fixed manually.)