Starting to build a db-only Geocoder() impl.
Ripping off the code for AddressGeocoder, but replacing all the Django model references with calls to the PostgisBlockSearcher.  In other words, starting to work back up to an interface that can be recognizably plugged back into a Django installation.  Also started adding support for the spelling correction element, which is (another) table in Django, but should probably be its own standalone interface here.