Fixing indenting, settled on class->file mapping.
(1) My vi was messing up the indent.  I've moved to editing with python-mode in emacs, and I've fixed (I think) all the indentation mistakes.  Sorry!

(2) Adam and I discussed where different classes should go.  This version has the "Searcher" classes in postgis.py (these are the classes that deal directly with the Postgis/Postgresql database), and the *Geocoder classes in djeocoder.py (these classes use the Searcher classes, and don't necessarily depend on the database directly).