Added LocatableResult, fixed PostgisBlockSearcher.
Started two new classes, based on LocatableResult, objects of which should be returned by the Searchers (instead of "raw tuples" directly from the database).

Fixed a bug in PostgisBlockSearcher, that was dropping all but the first result from the database (due to a bad indentation fix on my part).

Added some TODO and other documentation.

Starting moving the "contains_number" method to the BlockResult class.