Both Searcher classes result Result objects.
Updated PostgisBlockSearcher to return BlockResult objects, and updated
PostgisIntersectionSearcher to return IntersectionResult objects.

Debugged a line in the PostgisIntersectionSearcher's search() method, that was
not picking up the right 'street_b' arguments.

Separated out the wkt Point regex into its own method, postgis.parse_point()
(and created a PointParsingException that it throws on error).

Started adding more descriptive comments to the tops of the *Searcher classes.

Updated test.py with a test_PostgisIntersectionSearcher method, and updated
test_PostgisBlockSearcher to handle the new return type of PostgisBlockSearcher.search().