Debugging the AddressGeocoder.
Added in a few new exceptions which we hadn't carried over from the eblock code yet.

Trying to clean up and debug the actual code, so that the example in test.py works.

It's not working yet (as of this commit).  Still trying to figure out what happens, and why the geocoding complains,
when there are multiple blocks with the same name but different numbers, and only one of the numbers happens to match
(through the range) the number you've given it.